add timeout to pdf conversion in fulltext.py and ability to restart download_convert.py

+++ download_convert.py
@@ -4,7 +4,7 @@ import os
 from time import time
 from glob import glob
 from multiprocessing import Pool, cpu_count # Pool is used to create multiple processes
-from scientific_dataset_arxiv.fulltext import convert_directory_parallel
+from scientific_dataset_arxiv.fulltext import convert_directory_parallel, reextension
 from scientific_dataset_arxiv.config import start_year, end_year, max_pdfs_per_month
 
 
@@ -26,10 +26,9 @@ def create_folder(directory_path):
     if not os.path.exists(directory_path):
         os.makedirs(directory_path)
 
-## Function to download a folder from the bucket
 def download_folder_transfer_manager(bucket_name, bucket_folder_name, local_folder_path, workers=cpu_count(), max_results=max_pdfs_per_month):
     """
-    Downloads a folder from the bucket.
+    Downloads a folder from the bucket, skipping PDFs with corresponding TXT files.
 
     Args:
         bucket_name (str): The name of the bucket.
@@ -47,16 +46,48 @@ def download_folder_transfer_manager(bucket_name, bucket_folder_name, local_fold
 
     from google.cloud.storage import Client, transfer_manager
 
-    ## Create the folder if it doesn't exist
+    # Create the folder if it doesn't exist
     create_folder(local_folder_path)
 
-    ## Create an anonymous client for the bucket
+    # Create an anonymous client for the bucket
     storage_client = Client.create_anonymous_client()
 
-    ## Get the bucket and list the blobs
+    # Get the bucket and list the blobs
     bucket = storage_client.bucket(bucket_name)
 
-    blob_names = [blob.name for blob in bucket.list_blobs(max_results=max_results, prefix=bucket_folder_name)]
+    # Get list of existing TXT filenames
+    existing_txt_files = glob(f"{local_folder_path}/**/*.txt", recursive=True)
+
+    # Strip local_folder_path from each path, to match with blob names
+    existing_txt_files = [os.path.relpath(path, local_folder_path) for path in existing_txt_files]
+
+    # Normalize all paths in existing_txt_files and convert to a set for faster lookup
+    existing_txt_files = set(os.path.normpath(path) for path in existing_txt_files)
+
+    # Filter blobs to download (skip PDFs with corresponding TXT)
+    blobs_to_download = []
+    for blob in bucket.list_blobs(prefix=bucket_folder_name, max_results=max_results):
+
+        # Get blob name
+        blob_name = blob.name
+        
+        # Get the corresponding txt file name
+        txt_filename = reextension(blob_name, 'txt')
+
+        # Normalize the corresponding txt file name
+        txt_filename = os.path.normpath(txt_filename)
+
+        # Check if corresponding TXT file exists using set membership
+        # If the txt file does not exist, add the blob to the list of blobs to download
+        if txt_filename not in existing_txt_files:
+            blobs_to_download.append(blob)
+
+
+    blob_names = [blob.name for blob in blobs_to_download]
+
+    if not blob_names:
+        print(f"No new PDFs to download in {bucket_folder_name}, as all have corresponding TXT files.")
+        return
 
     results = transfer_manager.download_many_to_path(
         bucket, blob_names, destination_directory=local_folder_path, max_workers=workers, skip_if_exists=True
@@ -72,6 +103,7 @@ def download_folder_transfer_manager(bucket_name, bucket_folder_name, local_fold
             print("Downloaded {} to {}.".format(name, local_folder_path + name))
 
 
+
 ## Function to delete the original pdfs after they are converted to txt files
 def delete_pdfs_safe(directory_path):
     """
@@ -84,30 +116,31 @@ def delete_pdfs_safe(directory_path):
         None
     """
 
-    ## Get all pdf files
     pdf_files = glob(f"{directory_path}/**/*.pdf", recursive=True)
-    
-    ## Get all txt files
-    txt_files = glob(f"{directory_path}/**/*.txt", recursive=True)
-    
-    ## Convert to a set for faster searching
-    txt_files = set(txt_files)
-    
+    txt_files = set(glob(f"{directory_path}/**/*.txt", recursive=True))
 
-    ## Remove pdf only if there is a corresponding txt file
+    ## Check if there are any pdf files
+    if not pdf_files:  # Check if pdf_files is empty
+        print(f"No PDFs found to delete in: {directory_path}")
+        return
+
+    ## Check if there are any txt files
     for pdf in pdf_files:
-        
-        ## Get the pdf name
-        # pdf_name = pdf.split('/')[-1].split('.')[0] + '.' + pdf.split('/')[-1].split('.')[1]
-        
-        ## Get the txt name
-        txt_name = pdf.replace('.pdf', '.txt')
-        
-        ## Check if txt file exists
+
+        ## Get the txt file name
+        txt_name = reextension(pdf, 'txt')
+
+        ## Check if the txt file exists
         if txt_name in txt_files:
-            
-            ## Remove the pdf
-            os.remove(pdf)
+            try:
+                os.remove(pdf)
+                print(f"Deleted PDF: {pdf}")
+            except OSError as e:
+                print(f"Error deleting PDF: {pdf} - {e}")  # Log errors
+                continue
+        else:
+            print(f"No corresponding TXT file found for: {pdf}, hence will not delete")  # Handle missing TXT
+
 
 ## Creating a list for the year and month
 def create_yymm_list(start_year, end_year):
@@ -165,12 +198,15 @@ if __name__ == '__main__':
         local_folder_path = f'unprocessed_txts_{start_year}_to_{end_year}/{yymm}'
         
         ## Download all (max 10,000) the pdfs published on Arxiv in the year 20yy and month mm
+        print(f"Downloading PDFs for {yymm}.")
         download_folder_transfer_manager(bucket_name='arxiv-dataset', bucket_folder_name=f'arxiv/arxiv/pdf/{yymm}', local_folder_path=local_folder_path)
 
         ## Convert all the pdfs in the yymm directory to text
+        print(f"Converting PDFs to TXTs for {yymm}.")
         convert_directory_parallel(local_folder_path, cpu_count())
         
         ## Delete them pdfs if they have been converted to txts
+        print(f"Deleting PDFs for {yymm}.")
         delete_pdfs_safe(local_folder_path)
     
     ## Track time
